Added babel compilation + Removed barrel files + polyfils
