+++ packages/modul-components/.babelrc
@@ -4,8 +4,27 @@
             "@babel/preset-env",
             {
                 "useBuiltIns": "usage",
-                "corejs": 3
+                "corejs": 3,
+                "modules": "false"
+            }
+        ],
+        [
+            "@babel/preset-typescript",
+            {
+                "allowNamespaces": true
             }
         ]
+    ],
+    "plugins": [
+        [
+            "@babel/plugin-proposal-decorators",
+            {
+                "legacy": true
+            }
+        ],
+        [
+            "@babel/plugin-proposal-class-properties"
+        ],
+        "@babel/proposal-object-rest-spread"
     ]
 }
