+++ packages/modul-components/conf/webpack.lib.js
@@ -67,14 +67,6 @@ module.exports = {
                 use: ['style-loader',
                     'css-loader']
             },
-
-            {
-                test: /\.m?js$/,
-                exclude: /(node_modules)/,
-                use: {
-                    loader: 'babel-loader'
-                }
-            },
             {
                 test: /\.html$/,
                 loader: 'vue-template-loader',
