+++ packages/modul-components/package.json
@@ -7,11 +7,12 @@
     "private": false,
     "scripts": {
         "clean": "rimraf dist",
-        "build": "yarn run copyfiles && yarn run assetsToDist && tsc --project tsconfig.lib.json && yarn run buildlib && yarn run metagen",
-        "buildlib": "webpack --config conf/webpack.lib.js --colors",
+        "build": "yarn run copyfiles && yarn run assetsToDist && yarn run build:js && yarn run build:types && yarn run build:lib && yarn run metagen",
+        "build:types": "tsc --project tsconfig.lib.json --emitDeclarationOnly",
+        "build:js": "babel src --out-dir dist  --ignore \"src/**/*.d.ts\",\"src/**/*.spec.ts\",\"src/**/*.test.ts\" --extensions \".ts,.tsx\" --source-maps true",
+        "build:lib": "webpack --config conf/webpack.lib.js --colors",
         "copyfiles": "cpx \"./src/**/*.{json,html,scss,css}\" dist",
         "assetsToDist": "cpx \"./src/assets/**/*\" dist/assets",
-        "watch": "concurrently \"cpx \"./src/**/*.{json,html,scss,css}\" dist -w\" \"tsc --project tsconfig.lib.json -w\"  ",
         "start": "webpack --config conf/webpack.config.js --env.openshift && http-server \"./dist\" -p 8080 -a 0.0.0.0",
         "test": "jest --verbose",
         "test:update": "jest --verbose --silent --updateSnapshot",
@@ -36,6 +37,7 @@
         "axios": "^0.19.0",
         "cleave.js": "^1.5.0",
         "clipboard": "^2.0.4",
+        "core-js": "^3.4.2",
         "croppie": "^2.6.4",
         "css-element-queries": "^0.4.0",
         "enquire.js": "^2.1.6",
@@ -59,25 +61,26 @@
         "vue-router": "^3.0.1"
     },
     "devDependencies": {
+        "@babel/cli": "^7.7.4",
         "@babel/core": "^7.7.4",
+        "@babel/plugin-proposal-class-properties": "^7.7.4",
+        "@babel/plugin-proposal-decorators": "^7.7.4",
+        "@babel/plugin-proposal-object-rest-spread": "^7.7.4",
         "@babel/preset-env": "^7.7.4",
+        "@babel/preset-typescript": "^7.7.4",
         "@types/cleave.js": "^1.4.0",
         "@types/croppie": "^2.5.4",
         "@types/filesize": "^3.6.0",
         "@types/googlemaps": "^3.37.0",
+        "@types/jest": "^24.0.23",
         "@types/sprintf-js": "^1.1.0",
         "@types/webpack-env": "^1.13.6",
         "@vue/test-utils": "^1.0.0-beta.29",
         "autoprefixer": "^9.4.7",
-        "babel-loader": "^8.0.6",
-        "concurrently": "^5.0.0",
-        "core-js": "^3.4.2",
         "cpx": "^1.5.0",
         "css-loader": "^2.1.0",
         "file-gen-cli": "^1.2.2",
         "file-loader": "^3.0.1",
-        "fork-ts-checker-webpack-plugin": "^0.5.2",
-        "html-webpack-plugin": "^3.2.0",
         "jest": "^24.9.0",
         "jest-html-reporter": "^2.6.2",
         "jest-junit": "^8.0.0",
@@ -92,11 +95,7 @@
         "stylelint-scss": "^3.1.3",
         "stylelint-webpack-plugin": "^0.10.5",
         "svg-inline-loader": "^0.8.0",
-        "thread-loader": "^2.1.2",
         "ts-jest": "^24.1.0",
-        "ts-loader": "^5.3.3",
-        "uglifyjs-webpack-plugin": "^2.1.2",
-        "url-loader": "^1.1.2",
         "vue": "^2.6.10",
         "vue-class-component": "^6.1.1",
         "vue-property-decorator": "^6.0.0",
@@ -106,8 +105,7 @@
         "vue-template-loader": "^1.0.0",
         "webpack": "^4.41.2",
         "webpack-bundle-analyzer": "^3.6.0",
-        "webpack-cli": "^3.2.3",
-        "webpack-dev-server": "^3.1.14"
+        "webpack-cli": "^3.3.10"
     },
     "repository": {
         "type": "git",
