--- packages/modul-components/src/components/autocomplete/index.ts
@@ -1,2 +1,0 @@
-export { MAutoCompleteResult } from './autocomplete';
-
