--- packages/modul-components/src/components/table/index.ts
@@ -1,2 +1,0 @@
-export { MColumnSortDirection, MColumnTable, MColumnTextAlign, MTableSkin } from './table';
-
