--- packages/modul-components/src/components/toggle-buttons/index.ts
@@ -1,2 +1,0 @@
-export { MToggleButton, MToggleButtonSkin } from './toggle-buttons';
-
