--- packages/modul-components/src/components/tree/index.ts
@@ -1,2 +1,0 @@
-export { MCheckboxes, MSelectionMode, TreeNode } from './tree';
-
